mobile/icon_generator/requirements.txt
Pillow==9.5.0
rembg==2.0.50
numpy==1.24.3